+++ example/token.g.dart
@@ -39,6 +39,7 @@ class Token extends _i1.GeneratedContract {
     return (response[0] as BigInt);
   }
 
+  /// Returns a live stream of all Transfer events emitted by this contract.
   Stream<Transfer> transfer({_i1.BlockNum? fromBlock, _i1.BlockNum? toBlock}) {
     final event = self.event('Transfer');
     final filter = _i1.FilterOptions.events(