+++ lib/contracts/erc20.dart
@@ -1,0 +1,1 @@
+export '../src/generated/erc20.g.dart';